backend/pyproject.toml
[project]
name = "atlas"
version = "0.1.0"
description = "All-media Tracking & Library Access System"
authors = [{ name = "Andrew LeCody" }]
readme = "DOCS.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.14.1",
    "fastapi[standard]>=0.115.11,<0.116.0",
    "httpx>=0.28.1",
    "psycopg[binary]>=3.2.5",
    "pydantic-settings>=2.8.1",
    "pyjwt[crypto]>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.38",
    "sqlmodel>=0.0.23",
]

[dependency-groups]
dev = [
    "mypy>=1.15.0",
    "ruff>=0.9.5",
]

[project.scripts]
atlas-cli = "cli:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.mypy]
strict = true
exclude = ["venv", ".venv", "alembic"]
plugins = ["pydantic.mypy"]

[tool.ruff]
exclude = ["alembic"]
line-length = 100

[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # eradicate
    "ERA",
    # FastAPI
    "FAST",
]
